refactor: separation of Bubble Sort component with react context
- Added BubbleSort component to visualize the bubble sort algorithm.
- Introduced color constants for different sorting states (sorted, unsorted, marked for swap, marked for compare).
- Created MainContextProvider to manage audio playback and sorting state.
- Implemented useBubbleSort hook to handle sorting logic and state updates.
- Added helper function to convert number array to ElementType for sorting visualization.
- Updated home route to include MainContextProvider for context management.
