+++ app/constants.ts
@@ -1,0 +1,5 @@
+export const COLOR_COMPARE = "#38CC77"
+export const COLOR_SWAP = "#E21717"
+export const COLOR_UNSORTED = "#E03B8B"
+export const COLOR_SORTED = "#5A20CB"
+export const COLOR_SORT_T = "#F7CD2E"
